Move ast_builder back to utils.
