+++ aibolit/ast/__init__.py
@@ -1,3 +1,2 @@
 from aibolit.ast.ast import AST  # noqa: F401
 from aibolit.ast.ast_node_type import ASTNodeType  # noqa: F401
-from aibolit.ast.ast_builder import build_ast  # noqa: F401
