+++ aibolit/metrics/cognitiveC/cognitive_c.py
@@ -1,9 +1,11 @@
 from javalang.tree import IfStatement, SwitchStatement, ForStatement, WhileStatement
 from javalang.tree import DoStatement, CatchClause, BreakStatement, ContinueStatement
 from javalang.tree import TernaryExpression, BinaryOperation, MethodDeclaration, MethodInvocation
-from aibolit.ast import build_ast
 import javalang
 from typing import List, Any, Type
+
+from aibolit.utils.ast_builder import build_ast
+
 
 increment_for: List[Type] = [
     IfStatement,
