+++ aibolit/metrics/fanout/FanOut.py
@@ -1,4 +1,5 @@
-from aibolit.ast import build_ast, AST, ASTNodeType
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 from itertools import islice
 from aibolit.ast.java_package import JavaPackage
