+++ aibolit/metrics/lcom4/lcom4.py
@@ -21,7 +21,8 @@
 # SOFTWARE.
 
 import networkx as nx  # type: ignore
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
+
 from aibolit.utils.cohesiongraph import CohesionGraph
