+++ aibolit/metrics/mda/mda.py
@@ -1,4 +1,4 @@
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class MDAMetric:
