+++ aibolit/metrics/ncss/ncss.py
@@ -22,7 +22,8 @@
 
 from networkx import dfs_labeled_edges  # type: ignore
 
-from aibolit.ast import AST, ASTNodeType, build_ast
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 
 
 class NCSSMetric():
