+++ aibolit/metrics/number_variables/numVariables.py
@@ -1,4 +1,5 @@
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
+
 from javalang.tree import VariableDeclarator, MemberReference
