+++ aibolit/patterns/classic_setter/classic_setter.py
@@ -3,7 +3,7 @@ from typing import List
 import javalang
 
 from aibolit.types_decl import LineNumber
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class ClassicSetter:
