+++ aibolit/patterns/empty_rethrow/empty_rethrow.py
@@ -22,7 +22,7 @@
 
 import javalang
 
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class EmptyRethrow:
