+++ aibolit/patterns/force_type_casting_finder/force_type_casting_finder.py
@@ -22,7 +22,8 @@
 
 from typing import List
 
-from aibolit.ast import AST, ASTNodeType, build_ast
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 
 
 class ForceTypeCastingFinder:
