+++ aibolit/patterns/implements_multi/implements_multi.py
@@ -1,6 +1,6 @@
 import javalang
 
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class ImplementsMultiFinder:
