+++ aibolit/patterns/joined_validation/joined_validation.py
@@ -21,7 +21,9 @@
 # SOFTWARE.
 
 from typing import List, Tuple
-from aibolit.ast import AST, ASTNodeType, build_ast
+
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 
 
 class JoinedValidation:
