+++ aibolit/patterns/many_primary_ctors/many_primary_ctors.py
@@ -27,7 +27,7 @@ import javalang.parse
 import javalang.tree
 
 from aibolit.types_decl import LineNumber
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class ManyPrimaryCtors(object):
