+++ aibolit/patterns/method_chaining/method_chaining.py
@@ -3,7 +3,8 @@ import javalang
 import uuid
 from collections import defaultdict
 
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
+
 
 class MethodChainFind:
