+++ aibolit/patterns/method_siblings/method_siblings.py
@@ -21,7 +21,8 @@
 # SOFTWARE.
 from typing import List
 from aibolit.types_decl import LineNumber
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
+
 import javalang
 import re
