+++ aibolit/patterns/multiple_while/multiple_while.py
@@ -1,4 +1,5 @@
-from aibolit.ast import AST, ASTNodeType, build_ast
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 
 
 class MultipleWhile:
