+++ aibolit/patterns/protected_method/protected_method.py
@@ -1,6 +1,6 @@
 import javalang
 
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 
 
 class ProtectedMethod:
