+++ aibolit/patterns/return_null/return_null.py
@@ -21,7 +21,9 @@
 # SOFTWARE.
 
 from typing import List
-from aibolit.ast import AST, ASTNodeType, build_ast
+
+from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType
 
 
 class ReturnNull:
