+++ aibolit/patterns/string_concat/string_concat.py
@@ -3,7 +3,7 @@ from typing import Tuple, Dict, List
 import javalang
 
 from aibolit.types_decl import LineNumber
-from aibolit.ast import build_ast
+from aibolit.utils.ast_builder import build_ast
 from aibolit.utils.utils import RemoveComments
