+++ aibolit/patterns/var_siblings/var_siblings.py
@@ -1,7 +1,8 @@
 import javalang
 import re
-from aibolit.ast import build_ast
 import itertools
+
+from aibolit.utils.ast_builder import build_ast
 
 
 class VarSiblings:
